refactor: 클라이언트에서 데이터를 받아오는 방향으로 변경 후 에러처리
